Deallocation date functionality - change long types to ints for create allocation request

+++ SocialCareCaseViewerApi/V1/Boundary/Requests/CreateAllocationRequest.cs
@@ -10,10 +10,10 @@ namespace SocialCareCaseViewerApi.V1.Boundary.Requests
         public long MosaicId { get; set; }
 
         [JsonPropertyName("allocatedTeamId")]
-        public long AllocatedWorkerId { get; set; }
+        public int AllocatedWorkerId { get; set; }
 
         [JsonPropertyName("allocatedWorkerId")]
-        public long AllocatedTeamId { get; set; }
+        public int AllocatedTeamId { get; set; }
 
         [JsonPropertyName("createdBy")]
         public string CreatedBy { get; set; }
